rename: Basic Test의 폴더 구조를 맞춰줌.
